requirements-test.txt
-r requirements.txt
flake8
pytest
pytest-cov
pytest-django
requests-mock
